improve performance of quote-service

+++ client/js/controllers/chart.js
@@ -9,10 +9,15 @@ angular.module('StockSync')
             $scope.generate($scope.scalebyStart);
         });
         $scope.remove = function(symbol) {
-            QuoteService.remove(symbol);
-            $scope.generate($scope.scalebyStart);
-            stockSocket.emit('remove', {
-                symbol: symbol
-            });
+            if ($rootScope.symbols.length > 1) {
+                QuoteService.remove(symbol);
+                $scope.generate($scope.scalebyStart);
+                stockSocket.emit('remove', {
+                    symbol: symbol
+                });
+            }
+            else {
+                alert("The chart must have at least one symbol.");
+            }
         };
     }]);
